dockerfiles/x64/requirements.txt
spika==0.10.0
psutil==5.2.2
psycopg2==2.7.1
pymongo==3.4.0
python-consul==0.7.0
redis==2.10.5
requests==2.13.0
